Use augmented `RunTraceSpan` for extra data
