+++ pkg/coreapi/graph/models/models_gen.go
@@ -371,36 +371,6 @@ type RunStepInfo struct {
 
 func (RunStepInfo) IsStepInfo() {}
 
-type RunTraceSpan struct {
-	AppID          uuid.UUID          `json:"appID"`
-	FunctionID     uuid.UUID          `json:"functionID"`
-	RunID          ulid.ULID          `json:"runID"`
-	Run            *FunctionRun       `json:"run"`
-	SpanID         string             `json:"spanID"`
-	TraceID        string             `json:"traceID"`
-	Name           string             `json:"name"`
-	Status         RunTraceSpanStatus `json:"status"`
-	Attempts       *int               `json:"attempts,omitempty"`
-	Duration       *int               `json:"duration,omitempty"`
-	OutputID       *string            `json:"outputID,omitempty"`
-	QueuedAt       time.Time          `json:"queuedAt"`
-	StartedAt      *time.Time         `json:"startedAt,omitempty"`
-	EndedAt        *time.Time         `json:"endedAt,omitempty"`
-	ChildrenSpans  []*RunTraceSpan    `json:"childrenSpans"`
-	StepOp         *StepOp            `json:"stepOp,omitempty"`
-	StepID         *string            `json:"stepID,omitempty"`
-	StepInfo       StepInfo           `json:"stepInfo,omitempty"`
-	StepType       string             `json:"stepType"`
-	IsRoot         bool               `json:"isRoot"`
-	ParentSpanID   *string            `json:"parentSpanID,omitempty"`
-	ParentSpan     *RunTraceSpan      `json:"parentSpan,omitempty"`
-	IsUserland     bool               `json:"isUserland"`
-	UserlandSpan   *UserlandSpan      `json:"userlandSpan,omitempty"`
-	DebugRunID     *ulid.ULID         `json:"debugRunID,omitempty"`
-	DebugSessionID *ulid.ULID         `json:"debugSessionID,omitempty"`
-	DebugPaused    bool               `json:"debugPaused"`
-}
-
 type RunTraceSpanOutput struct {
 	Input *string    `json:"input,omitempty"`
 	Data  *string    `json:"data,omitempty"`
